feat(pip_setup): Separate testing of public and "private" API
