+++ test/manager/pip_setup/extract.spec.ts
@@ -1,64 +1,15 @@
-import { readFileSync } from 'fs';
-import { file as _file } from 'tmp-promise';
-import { relative } from 'path';
 import { exec } from '../../../lib/util/exec';
+
 import {
-  extractPackageFile,
   parsePythonVersion,
   getPythonAlias,
   pythonVersions,
 } from '../../../lib/manager/pip_setup/extract';
 
-const packageFile = 'test/manager/pip_setup/_fixtures/setup.py';
-const content = readFileSync(packageFile, 'utf8');
-const config = {
-  localDir: '.',
-};
-
-async function tmpFile() {
-  const file = await _file({ postfix: '.py' });
-  return relative('.', file.path);
-}
-
 describe('lib/manager/pip_setup/extract', () => {
   beforeEach(() => {
     jest.resetModules();
   });
-  describe('extractPackageFile()', () => {
-    it('returns found deps', async () => {
-      expect(
-        await extractPackageFile(content, packageFile, config)
-      ).toMatchSnapshot();
-    });
-    it('should return null for invalid file', async () => {
-      expect(
-        await extractPackageFile('raise Exception()', await tmpFile(), config)
-      ).toBeNull();
-    });
-    it('should return null for no deps file', async () => {
-      expect(
-        await extractPackageFile(
-          'from setuptools import setup\nsetup()',
-          await tmpFile(),
-          config
-        )
-      ).toBeNull();
-    });
-    it('catches error', async () => {
-      const fExec = jest.fn(() => {
-        throw new Error('No such file or directory');
-      });
-      jest.doMock('../../../lib/util/exec', () => {
-        return {
-          exec: fExec,
-        };
-      });
-      const m = require('../../../lib/manager/pip_setup/extract');
-      await m.extractPackageFile(content, packageFile, config);
-      expect(fExec).toHaveBeenCalledTimes(4);
-    });
-  });
-
   describe('parsePythonVersion', () => {
     it('returns major and minor version numbers', () => {
       expect(parsePythonVersion('Python 2.7.15rc1')).toEqual([2, 7]);
@@ -104,31 +55,4 @@ describe('lib/manager/pip_setup/extract', () => {
       expect(isMockInstalled).toBe(true);
     });
   });
-  /*
-  describe('extractSetupFile()', () => {
-    it('should return parsed setup() call', async () => {
-      expect(
-        await extractSetupFile(content, packageFile, config)
-      ).toMatchSnapshot();
-    });
-    it('should support setuptools', async () => {
-      expect(
-        await extractSetupFile(
-          'from setuptools import setup\nsetup(name="talisker")\n',
-          await tmpFile(),
-          config
-        )
-      ).toEqual({ name: 'talisker' });
-    });
-    it('should support distutils.core', async () => {
-      expect(
-        await extractSetupFile(
-          'from distutils.core import setup\nsetup(name="talisker")\n',
-          await tmpFile(),
-          config
-        )
-      ).toEqual({ name: 'talisker' });
-    });
-  });
-  */
 });
